+++ pyveg/src/process_satellite_data.py
